Start switching to node-mysql
